@@ -37,17 +37,17 @@
     "@swc/core": "^1.9.3",
     "@swc/jest": "^0.2.37",
     "@types/node": "^22.10.1",
-    "@types/react": "^18.3.10",
+    "@types/react": "^19.0.1",
     "concurrently": "^9.1.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "size-limit": "^11.1.6",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.2"
   },
   "peerDependencies": {
-    "react": "^18"
+    "react": "^18 || ^19"
   },
   "size-limit": [
     {
