@@ -42,15 +42,15 @@
     "@size-limit/webpack": "^11.1.6",
     "@swc/cli": "^0.5.2",
     "@swc/core": "^1.9.3",
-    "@types/react": "^18.3.3",
+    "@types/react": "^19.0.1",
     "@vkontakte/icons-scripts": "workspace:*",
     "css-loader": "^7.1.2",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.6.3",
     "pre-commit": "^1.2.2",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "react-color": "^2.18.0",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "size-limit": "^11.1.6",
     "style-loader": "^4.0.0",
     "swc-loader": "0.2.6",
@@ -58,7 +58,7 @@
     "webpack": "^5.96.1"
   },
   "peerDependencies": {
-    "react": "^18"
+    "react": "^18 || ^19"
   },
   "size-limit": [
     {
