@@ -1,5 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import * as React from 'react';
+import { createRoot } from 'react-dom/client';
 import { aliases } from './aliases.js';
 import { HuePicker as Hue } from 'react-color';
 import './docs.css';
@@ -183,4 +183,5 @@ class Docs extends React.PureComponent {
   }
 }
 
-ReactDOM.render(<Docs />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+root.render(<Docs />);
